make "list" command show more info by default
The original command just showed a simple list, but this was pretty
useless: The person listing the services is doing so to get an idea of
what's available to run, and the person will likely never, ever want to
run more than one service by piping the result of the "list" command to
another command.

Now, the "list" command shows a `$summary` from the container file and
the `=head1 NAME` section from the POD of the service, which should
include a short abstract about the runnable module itself.