add summary attribute for runnable services
The summary attribute allows for better control over what shows up in
the `beam list` command when we have a highly configurable runnable
module.

This still does not instantiate the object, which means that we can't
provide a good default inside the class itself. This is a problem that
should be addressed, but instantiating a module means loading all its
dependencies (like database connections, log filehandles, and other
such) that could mean a lot of work and/or DoS conditions. We should
figure out a way to fix this...